Update to reflect latest changes
Also lowered the chance of an ordering issue, and improved documentation

+++ config.go
@@ -2,8 +2,8 @@ package testimpl
 
 // Config is the configuration structure of the Docker connector.
 type Config struct {
-	// The number of seconds it should wait while deploying to mimic a real deployer.
-	DeployTime          int32 `json:"deploy_time"`
+	// The number of milliseconds seconds it should wait while deploying to mimic a real deployer.
+	DeployTime          int64 `json:"deploy_time"`
 	DeploySucceed       bool  `json:"deploy_succeed"`
 	DisablePluginWrites bool  `json:"disable_plugin_writes"`
 }
